Run tests in parallel forks
